+++ src/widgets/header/ui/theme_switcher/index.tsx
